pyproject.toml
[project]
name = "py311-chzzk-auto-recorder"
version = "0.1.0"
description = "치지직 자동 녹화 시스템"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "httpx",
    "python-dotenv", 
    "structlog",
    "streamlink",  # HLS 스트림 다운로드용 (FFmpeg 기반)
]
requires-python = ">= 3.11"

[project.optional-dependencies]
monitor = [
    "beautifulsoup4",
    "lxml",
]
recorder = [
    # ffmpeg는 시스템 패키지로 설치
]
dev = [
    "pytest",
    "pytest-asyncio",
    "black",
    "ruff",
    "mypy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chzzk_recorder"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
